requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==6.0.5
djangorestframework==3.17.1
gunicorn==26.0.0
idna==3.15
packaging==26.2
python-dotenv==1.2.2
requests==2.34.1
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
whitenoise==6.12.0
